% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{begin}
\alias{begin}
\title{Record Running Time of a Program}
\usage{
begin()
}
\value{
Several objects loaded to the global environment
\itemize{
  \item begin_time - Starting time of the program
  \item end_time - Ending time of the program
  \item runtime - Numeric value containing elapsed seconds
}
}
\description{
`begin()` starts running the stopwatch and `end()` stops the stopwatch to record run time.
}
\examples{
begin()
Sys.sleep(10)
end()
}
