% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_packages}
\alias{load_packages}
\title{Install & Load Packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{a character vector containing package names}
}
\description{
Installs packages that are not installed and loads the already installed packages.
}
\examples{
load_packages(packages = c("tidyverse","readxl"))
}
