% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultToDummy.R
\name{MultToDummy}
\alias{MultToDummy}
\title{Split Multiple Response Variables into Dummies}
\usage{
MultToDummy(x, splitby = ";", bind = TRUE)
}
\arguments{
\item{x}{Character vector of multiple responses}

\item{splitby}{Character string used to split the multiple responses}

\item{bind}{Logical. If bind is TRUE, the original variable is returned along with the dummy variables. If bind is FALSE, only the dummy variables are returned. Default is TRUE.}
}
\value{
A data frame with the original variable and the dummy variables, or just the dummy variables if bind is FALSE.
}
\description{
Split Multiple Response Variables into Dummies
}
\examples{
# Example for split by semicolon:
MultToDummy(x, splitby = ";")

# Example for split by capital letters:
MultToDummy(x, splitby = "(?<=.)(?=[A-Z])")

}
