% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{seconds_to_time_string}
\alias{seconds_to_time_string}
\title{Seconds to Time String Conversion}
\usage{
seconds_to_time_string(seconds)
}
\arguments{
\item{seconds}{numeric value containing seconds}
}
\value{
seconds in time string
}
\description{
Converts a numeric value to time string
}
\examples{
seconds_to_time_string(67)
seconds_to_time_string(5*60*60+9*60+58)
seconds_to_time_string(45*24*60*60+13*60*60+15*60+25)
}
