% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{total_runtime}
\alias{total_runtime}
\title{Calculate Total Run-time of a Program}
\usage{
total_runtime(runtime_file = "RUNTIME.xlsx")
}
\arguments{
\item{runtime_file}{character value containing the file path to the Run-time Excel file. Default to "RUNTIME.xlsx"}
}
\description{
This function calculates total run-time of a program from the run-time excel file created using the function `write_runtime()`.
}
\examples{
begin()
Sys.sleep(10)
end()
write_runtime("Task 1", runtime, "Duration.xlsx")
begin()
Sys.sleep(15)
end()
write_runtime("Task 2", runtime, "Duration.xlsx")
total_runtime("Duration.xlsx")
# Data	  Runtime
# Task 1	10
# Task 2	15
# Total	25 Seconds
}
