% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{write_runtime}
\alias{write_runtime}
\title{Record Run-time to an Excel File}
\usage{
write_runtime(dataname, runtime = runtime, runtime_file = "RUNTIME.xlsx")
}
\arguments{
\item{dataname}{character value containing the name of the program}

\item{runtime}{numeric value containing the seconds took to run the program. Default to `runtime` returned by `end_time()`}

\item{runtime_file}{character value containing the file path to the Excel file. Default to "RUNTIME.xlsx"}
}
\description{
Records duration of a program to an Excel file. Plays a sound when task is successful.
}
\examples{
begin()
Sys.sleep(10)
end()
write_runtime("Task 1", runtime, "Duration.xlsx")
}
